Whats new: 1.Timeshift and Realworkhour able to remove preapended employees 2.Delete functionality works almost anywhere, complete

+++ WebApplication/wwwroot/js/Custom/DatatableDeleteButton.js
@@ -1,11 +1,37 @@
 ﻿//Delete action in datatable
-$(document).on('click', '.DatatableDeleteButton', (element) => {
-    $.ajax({
-        type: "DELETE",
-        url: element.target.getAttribute('urlattr'),
-        dataType: "json",
-        success: function (response) {
-            console.log('mmmkeyyyyy');
+$(document).on('click', '.DatatableDeleteButton', element =>
+    swal({
+        title: 'Προσοχή!',
+        text: 'Είσαστε σίγουρος οτι θέλετε να διαγράψετε την οντώτητα;',
+        icon: 'warning',
+        closeOnClickOutside: false,
+        buttons: {
+            cancel: {
+                text: "Απόριψη",
+                value: null,
+                visible: true,
+                className: "",
+                closeModal: true,
+            },
+            accept: {
+                text: "Αποδοχή",
+                value: true,
+                visible: true,
+                className: "",
+                closeModal: true,
+            }
         }
-    });
-})
+    }).then((value) => {
+        if (value === true)
+            $.ajax({
+                type: "DELETE",
+                url: element.target.getAttribute('urlattr'),
+                dataType: "json",
+                success: (response) => {
+                    console.log(response);
+                    swal(response.responseTitle, response.responseBody, (response.isSuccessful ? 'success' : 'error'));
+                }
+            }).then(() =>
+                [...document.getElementsByClassName("table")]
+                    .forEach(table => $(table).DataTable().ajax.reload()));
+    }));
